feature/LoginAndSignUp/BE

+++ backend/src/main/java/com/ssafy/niceage/Domain/Senior_Citizen_Center/Senior_Citizen_Center.java
@@ -14,7 +14,7 @@ import lombok.Data;
 public class Senior_Citizen_Center {
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column (name = "senior_id")
-	private int seniorId;
+	private Long seniorId;
 	
 	@OneToMany(mappedBy = "senior", cascade = CascadeType.ALL, orphanRemoval =true)
 	List<Enter> enters = new ArrayList<Enter>();
